Preview: get file data from File if not in FileList
The file data requested may not be located in the FileList. This
happens if, for example, the user is visiting the appraisal tab from
a link to the preview pane which contains a file UUID; the FileList
will be empty in this case.

This still fetches the file from the FileList where possible, but will
fall back to fetching it from the server otherwise.