Vendor angular-charts
Angular-charts is not published to npm, only bower. The git repository
does contain a package.json, allowing it to be installed via npm using
the git repo's URL; however, this introduces several buildtime deps in
order to produces a distribution-ready .js file. To simplify things,
instead just vendor the specific version that we use.