tree-droppable: adjust tolerance
The default tolerance is 'intersect', which is documented as firing if
the "Draggable overlaps the droppable at least 50% in both directions".
This isn't nearly good enough for the elements we're dragging; it's
more intuitive for the user if we instead consider the mouse cursor to
be the focal point of dragging.

See the docs: http://api.jqueryui.com/droppable/