Remove CSS that comes from Archivematica
This CSS is copied from Archivematica. Not all of it is used; the parts
that are used are used in the directory picker component.

For now, remove these so Archivematica's global definitions provide them.
In the future, it'd be good to possibly pare these down to only the
subset the embeddable component uses so that can be used in more than
one project without them stomping over each other.